+++ superduperdb/backends/mongodb/query.py
@@ -23,6 +23,7 @@ from superduperdb.backends.base.query import (
 from superduperdb.base.cursor import SuperDuperCursor
 from superduperdb.base.document import Document
 from superduperdb.components.schema import Schema
+from superduperdb.misc.auto_schema import infer_schema
 from superduperdb.misc.files import load_uris
 
 SCHEMA_KEY = '_schema'
@@ -955,6 +956,17 @@ class Collection(TableOrCollection):
                 db.execute(
                     collection.insert_many([Document(**doc) for doc in bulk_writes])
                 )
+
+    @staticmethod
+    def infer_schema(data: t.Mapping[str, t.Any], identifier: t.Optional[str] = None):
+        """
+        Infer a schema from a given data object
+
+        :param data: The data object
+        :param identifier: The identifier for the schema, if None, it will be generated
+        :return: The inferred schema
+        """
+        return infer_schema(data, identifier)
 
 
 def _get_decode_function(db) -> t.Callable[[t.Any], t.Any]: